+++ core/security/digest/src/test/kotlin/org/http4k/filter/DigestAuthTest.kt
@@ -11,6 +11,7 @@ import org.http4k.core.then
 import org.http4k.hamkrest.hasHeader
 import org.http4k.hamkrest.hasStatus
 import org.http4k.security.Nonce.Companion.SECURE_NONCE
+import org.http4k.security.digest.DigestAlgorithm.MD5
 import org.http4k.security.digest.DigestMode.Proxy
 import org.http4k.security.digest.DigestMode.Standard
 import org.http4k.security.digest.Qop.Auth
@@ -24,7 +25,7 @@ class DigestAuthTest {
 
     @Test
     fun `invalid credentials`() {
-        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, listOf(Auth), Standard, SECURE_NONCE, { true })
+        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, listOf(Auth), Standard, SECURE_NONCE, { true }, MD5)
             .then { Response(OK) }
         val response = ClientFilters.DigestAuth(Credentials("admin", "hunter2")).then(handler)(Request(GET, "/"))
 
@@ -35,7 +36,7 @@ class DigestAuthTest {
     @Test
     fun `valid credentials with Auth Qop`() {
         val handler =
-            ServerFilters.DigestAuth(REALM, passwordLookup, listOf(Auth), Standard, SECURE_NONCE, { true })
+            ServerFilters.DigestAuth(REALM, passwordLookup, listOf(Auth), Standard, SECURE_NONCE, { true }, MD5)
                 .then { Response(OK) }
         val response = ClientFilters.DigestAuth(Credentials("admin", "password")).then(handler)(Request(GET, "/"))
 
@@ -55,7 +56,7 @@ class DigestAuthTest {
 
     @Test
     fun `valid credentials with no qop`() {
-        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, emptyList(), Standard, SECURE_NONCE, { true })
+        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, emptyList(), Standard, SECURE_NONCE, { true }, MD5)
             .then { Response(OK) }
         val response = ClientFilters.DigestAuth(Credentials("admin", "password"))
             .then(handler)(Request(GET, "/"))
@@ -66,7 +67,7 @@ class DigestAuthTest {
 
     @Test
     fun `valid credentials with full uri`() {
-        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, emptyList(), Standard, SECURE_NONCE, { true })
+        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, emptyList(), Standard, SECURE_NONCE, { true }, MD5)
             .then { Response(OK) }
         val response = ClientFilters.DigestAuth(Credentials("admin", "password"))
             .then(handler)(Request(GET, "http://server:80/path?foo=bar"))
@@ -77,7 +78,7 @@ class DigestAuthTest {
 
     @Test
     fun `valid credentials in proxy mode`() {
-        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, listOf(Auth), Proxy, SECURE_NONCE, { true })
+        val handler = ServerFilters.DigestAuth(REALM, passwordLookup, listOf(Auth), Proxy, SECURE_NONCE, { true }, MD5)
             .then { Response(OK) }
         val response = ClientFilters.DigestAuth(Credentials("admin", "password"), digestMode = Proxy)
             .then(handler)(Request(GET, "/"))